refactor!: the collider shape discriminant has one definition
pkg/spatial declared `ShapeCircle/ShapeRect uint8` and pkg/component declared
`ShapeKind` with values deliberately equal to them, kept in step by a comment
that said so. Two enums agreeing by convention is the arrangement phase 4 is
about to lean on hardest: the broad phase reads Shape to pick a narrow-phase
routine, and 4b adds a third discriminant.

pkg/spatial now reads component.ShapeKind directly. The import is acyclic —
`go list -f '{{.Imports}}' ./pkg/component` is `[math]`, so pkg/component
cannot reach back — and this edge is needed anyway for the Vec3 query centre
later in the phase.

The facade drops ShapeCircle/ShapeRect entirely rather than aliasing them.
A sphere IS a circle in a 2D profile, so the forward names are the only names
worth having, and an alias would leave the 2D spelling looking like a distinct
concept.

The rationale comment moves with the constraint rather than being deleted: the
hazard was never the aliasing, it was that a third shape added in one place
fails SILENTLY in two — spatial's narrow phase falls through to the OBB
routine for an unrecognised value while the raycast skips it. That is now
recorded where the enum lives.

No behaviour change; the values are identical. `just schema-check` green (Shape
is a wire byte and it did not move), `just lint-no-ark` green, examples/space
green.

+++ spatial_system.go
@@ -58,7 +58,7 @@ func (s *SpatialSystem) Update(dt float32) {
 			Width:  b.Col.Width,
 			Height: b.Col.Height,
 			Layer:  b.Col.Layer,
-			Shape:  uint8(b.Col.Shape),
+			Shape:  b.Col.Shape,
 		}
 		if b.Rot != nil {
 			entry.Rotation = b.Rot.Yaw()
